requirements.txt
langchain==0.2.12
faiss-cpu==1.8.0
sentence-transformers==2.7.0
transformers==4.43.3
torch>=2.0
openai==1.40.0
pypdf==4.2.0
python-docx==1.1.0
unstructured==0.14.9
tiktoken==0.7.0

